Conflicts:
	fusinvinventory/front/popup.php
	fusinvinventory/install/update_232_240.php
	fusinvinventory/install/update_240_0.80+1.1.php
	fusinvinventory/locales/de_DE.php
	fusinvinventory/locales/en_GB.php
	fusinvinventory/locales/es_ES.php
	fusinvinventory/locales/et_EE.php
	fusinvinventory/locales/fr_FR.php
	fusinvinventory/locales/it_IT.php
	fusinvinventory/locales/pl_PL.php
	fusinvinventory/locales/pt_BR.php
	fusinvinventory/locales/ru_RU.php
	fusinvinventory/locales/tr_TR.php
	fusioninventory/b/esx/index.php
	fusioninventory/front/communication.php
	fusioninventory/hook.php
	fusioninventory/inc/agent.class.php
	fusioninventory/inc/communication.class.php
	fusioninventory/inc/config.class.php
	fusioninventory/inc/inventorycomputeresx.class.php
	fusioninventory/inc/inventorycomputerinventory.class.php
	fusioninventory/inc/inventorycomputerlibhook.class.php
	fusioninventory/inc/taskjobstatus.class.php
	fusioninventory/inc/unknowndevice.class.php
	fusioninventory/install/update.php
	fusioninventory/install/update_232_240.php
	fusioninventory/locales/de_DE.php
	fusioninventory/phpunit/InventoryComputer/AllTests.php
	fusioninventory/setup.php
	fusioninventory/tools/phpunit/InventoryLocal.php

+++ fusioninventory/inc/inventorycomputerimport_monitor.class.php
@@ -3,7 +3,7 @@
 /*
    ------------------------------------------------------------------------
    FusionInventory
-   Copyright (C) 2010-2011 by the FusionInventory Development Team.
+   Copyright (C) 2010-2012 by the FusionInventory Development Team.
 
    http://www.fusioninventory.org/   http://forge.fusioninventory.org/
    ------------------------------------------------------------------------
@@ -30,7 +30,7 @@
    @package   FusionInventory
    @author    David Durieux
    @co-author 
-   @copyright Copyright (c) 2010-2011 FusionInventory team
+   @copyright Copyright (c) 2010-2012 FusionInventory team
    @license   AGPL License 3.0 or (at your option) any later version
               http://www.gnu.org/licenses/agpl-3.0-standalone.html
    @link      http://www.fusioninventory.org/
@@ -169,6 +169,14 @@ class PluginFusioninventoryInventoryComputerImport_Monitor extends CommonDBTM {
       if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
          $array['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
       }
+      if ($type == "add") { // Case where have same monitor xx times
+         $a_find = $Computer_Item->find("`computers_id`='".$items_id."'
+            AND `itemtype` = 'Monitor'
+            AND `items_id`='".$monitor_id."'");
+         if (count($a_find) > 0) {
+            return;
+         }
+      }
       $devID = $Computer_Item->add($array);
       return $devID;
    }
@@ -189,15 +197,16 @@ class PluginFusioninventoryInventoryComputerImport_Monitor extends CommonDBTM {
       if ($pfConfig->getValue($_SESSION["plugin_fusinvinventory_moduleid"],
                  "import_monitor", 'inventory') != '0') {
          $Computer_Item = new Computer_Item();
-         $Computer_Item->getFromDB($items_id);
-         if ($Computer_Item->fields['computers_id'] == $idmachine) {
-            $input = array();
-            $input['id'] = $items_id;
-            $input['itemtype'] = 'Monitor';
-            if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
-               $input['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
+         if ($Computer_Item->getFromDB($items_id)) {
+            if ($Computer_Item->fields['computers_id'] == $idmachine) {
+               $input = array();
+               $input['id'] = $items_id;
+               $input['itemtype'] = 'Monitor';
+               if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
+                  $input['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
+               }
+               $Computer_Item->delete($input, 0, $_SESSION["plugin_fusinvinventory_history_add"]);
             }
-            $Computer_Item->delete($input, 0, $_SESSION["plugin_fusinvinventory_history_add"]);
          }
       }
    }
