+++ fusioninventory/inc/inventorycomputerimport_peripheral.class.php
@@ -3,7 +3,7 @@
 /*
    ------------------------------------------------------------------------
    FusionInventory
-   Copyright (C) 2010-2011 by the FusionInventory Development Team.
+   Copyright (C) 2010-2012 by the FusionInventory Development Team.
 
    http://www.fusioninventory.org/   http://forge.fusioninventory.org/
    ------------------------------------------------------------------------
@@ -30,7 +30,7 @@
    @package   FusionInventory
    @author    David Durieux
    @co-author 
-   @copyright Copyright (c) 2010-2011 FusionInventory team
+   @copyright Copyright (c) 2010-2012 FusionInventory team
    @license   AGPL License 3.0 or (at your option) any later version
               http://www.gnu.org/licenses/agpl-3.0-standalone.html
    @link      http://www.fusioninventory.org/
@@ -176,6 +176,14 @@ class PluginFusioninventoryInventoryComputerImport_Peripheral extends CommonDBTM
       if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
          $array['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
       }
+      if ($type == "add") { // Case where have same peripheral xx times
+         $a_find = $Computer_Item->find("`computers_id`='".$items_id."'
+            AND `itemtype` = 'Peripheral'
+            AND `items_id`='".$peripheral_id."'");
+         if (count($a_find) > 0) {
+            return;
+         }
+      }
       $devID = $Computer_Item->add($array);
       return $devID;
    }
@@ -196,15 +204,16 @@ class PluginFusioninventoryInventoryComputerImport_Peripheral extends CommonDBTM
       if ($pfConfig->getValue($_SESSION["plugin_fusinvinventory_moduleid"],
                  "import_peripheral", 'inventory') != '0') {
          $Computer_Item = new Computer_Item();
-         $Computer_Item->getFromDB($items_id);
-         if ($Computer_Item->fields['computers_id'] == $idmachine) {
-            $input = array();
-            $input['id'] = $items_id;
-//            $input['itemtype'] = "Peripheral";
-            if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
-               $input['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
+         if ($Computer_Item->getFromDB($items_id)) {
+            if ($Computer_Item->fields['computers_id'] == $idmachine) {
+               $input = array();
+               $input['id'] = $items_id;
+   //            $input['itemtype'] = "Peripheral";
+               if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
+                  $input['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
+               }
+               $Computer_Item->delete($input, 0, $_SESSION["plugin_fusinvinventory_history_add"]);
             }
-            $Computer_Item->delete($input, 0, $_SESSION["plugin_fusinvinventory_history_add"]);
          }
       }
    }
